move citation parser to section
